src/Python/unstructure/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.5
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
effdet==0.4.1
email_validator==2.1.1
emoji==2.11.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.14.0
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.0.347
langchain-core==0.0.11
langdetect==1.0.9
langsmith==0.0.92
layoutparser==0.3.4
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.17.3
openai==1.28.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
ordered-set==4.1.0
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.15.1
pillow==10.3.0
pillow_heif==0.16.0
portalocker==2.8.2
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.29.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.40.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
unstructured==0.13.7
unstructured-client==0.22.0
unstructured-inference==0.7.31
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.18.1
